Add authentication to the GraphQL connection
We're using the apollo-boost version of the library so all we need to do
is drop a headers config into the connection including an auth header of
a Bearer token containing our Personal Access Token.

Run the command as `TOKEN=P_A_TOKEN_HERE ./main.js` and it'll slurp the
token out of the environment for you.