Finish label parsing and scoring
We now have a full set of parsers for the most common kinds of
labels. We can parse single values and pretty much any combination
of numbers, punctuation and the words story and points. We're also
doing the map-reduce right to get down to a totalled amount of
points per column. There's also a lot of code comments.